requirements.txt
asgiref==3.8.1
attrs==25.3.0
cattrs==24.1.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver==2.24.1
cloudinary==1.43.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
playwright==1.51.0
psycopg2-binary==2.9.10
pycparser==2.22
pyee==12.1.1
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.0
qrcode==8.0
requests==2.32.3
requests-cache==1.2.1
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.3.0
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
